+++ source_gen/source_gen/CHANGELOG.md
@@ -78,7 +78,7 @@ findTokenField(DartObject o) {
   old behavior for precise checks.
 * `TypeChecker#annotations...`-methods now throw a `StateError` if one or more
   annotations on an element are not resolvable. This is usually a sign of a
-  mispelling, missing import, or missing dependency.
+  misspelling, missing import, or missing dependency.
 * Added `TypeChecker.any`, which delegates to multiple other `TypeChecker`
   implementations when making a type check.